Updated camera-constrain.js to prevent vibration
Added additional constraint to camera constrain's limitToWorld function to handle the case where one or more axes of the view-port are larger than the world bounds. In that case the world is centered in the viewport on each of the offending axis.

This replaces the current behavior of the camera vibrating like mad as it bounces between trying to satisfy constraints on each side.